Fix warning if large file support is enabled

+++ src/gmdb2/info.c
@@ -81,11 +81,11 @@ MdbCatalogEntry *entry = mdb_get_catalogentry_by_name(mdb, "SummaryInfo");
 	gtk_label_set_text(GTK_LABEL(label), mdb->f->db_key ? "Yes" : "No");	
 
 	assert( fstat(mdb->f->fd, &st)!=-1 );
-	sprintf(tmpstr, "%ld K", st.st_size/1024);
+	sprintf(tmpstr, "%zd K", (size_t)(st.st_size/1024));
 	label = glade_xml_get_widget (propswin_xml, "props_filesize");
 	gtk_label_set_text(GTK_LABEL(label), tmpstr);	
 		
-	sprintf(tmpstr, "%ld", st.st_size / mdb->fmt->pg_size);
+	sprintf(tmpstr, "%zd", (size_t)(st.st_size / mdb->fmt->pg_size));
 	label = glade_xml_get_widget (propswin_xml, "props_numpages");
 	gtk_label_set_text(GTK_LABEL(label), tmpstr);
